requirements.txt
fastapi==0.110.3
uvicorn==0.30.5
httpx==0.27.2
pydantic==2.8.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-api-python-client==2.143.0
google-cloud-videointelligence==2.12.0
yt-dlp==2025.07.21
openai==1.51.0
pytest==8.3.2
